# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Biskora Cookies API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ingredient {
  id: number;
  productId: number;
  name: string;
  isInStock: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  description: string;
  imageUrl: string;
  highlights: string[];
  isAvailable: boolean;
  price: number;
  ingredients?: Ingredient[];
  createdAt: string;
  updatedAt: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  new: 'new',
  read: 'read',
  responded: 'responded',
  completed: 'completed',
} as const;

export interface Order {
  id: number;
  name: string;
  email: string;
  phone: string;
  message: string;
  orderType: string;
  status: OrderStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderInput {
  name: string;
  email: string;
  phone: string;
  message: string;
  orderType: string;
}

export interface CreateProductInput {
  name: string;
  category: string;
  description: string;
  imageUrl: string;
  highlights: string[];
  isAvailable: boolean;
  price: number;
}

export interface UpdateProductInput {
  name?: string;
  category?: string;
  description?: string;
  imageUrl?: string;
  highlights?: string[];
  isAvailable?: boolean;
  price?: number;
}

export interface CreateIngredientInput {
  productId: number;
  name: string;
  isInStock: boolean;
}

export interface UpdateIngredientInput {
  name?: string;
  isInStock?: boolean;
}

export type UpdateOrderInputStatus = typeof UpdateOrderInputStatus[keyof typeof UpdateOrderInputStatus];


export const UpdateOrderInputStatus = {
  new: 'new',
  read: 'read',
  responded: 'responded',
  completed: 'completed',
} as const;

export interface UpdateOrderInput {
  status: UpdateOrderInputStatus;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export interface AdminMe {
  email: string;
  role: string;
}

export interface AdminStats {
  totalOrders: number;
  newOrders: number;
  totalProducts: number;
  outOfStockIngredients: number;
}

export type GetProductsParams = {
category?: string;
};